requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
spacy==3.7.2
PyPDF2==3.0.1
python-docx==1.1.0
requests==2.31.0
beautifulsoup4==4.12.2
pydantic==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

